+++ adapters/repos/db/vector/hnsw/condensor_integration_test.go
@@ -203,7 +203,7 @@ func dumpIndexFromCommitLog(t *testing.T, fileName string) {
 	index := &hnsw{
 		nodes:               res.Nodes,
 		currentMaximumLayer: int(res.Level),
-		entryPointID:        int64(res.Entrypoint),
+		entryPointID:        res.Entrypoint,
 		tombstones:          res.Tombstones,
 	}
